Update start.sh
Don't execute git related functions when no remote repo is provided.

+++ start.sh
@@ -5,21 +5,27 @@ JVM_RAM="${SERVER_RAM}"
 
 cd /home/container
 
-if [ ! -d "$REPO_DIR" ]; then
-	echo -e "Git Directory does not exist. Cloning git repo..."
-	mkdir -p "$REPO_DIR"
-	ssh-agent bash -c "ssh-add /home/container/ssh/id-rsa; git clone -b ${GITHUB_REPO_BRANCH} --single-branch ${GITHUB_REPO_SSH} $REPO_DIR"
-fi
+if [ -n "${GITHUB_REPO_SSH}" ]; then
+	if [ ! -d "$REPO_DIR" ]; then
+		echo -e "Git Directory does not exist. Cloning git repo..."
+		mkdir -p "$REPO_DIR"
+		ssh-agent bash -c "ssh-add /home/container/ssh/id-rsa; git clone -b ${GITHUB_REPO_BRANCH} --single-branch ${GITHUB_REPO_SSH} $REPO_DIR"
+	fi
 
-# Start git-monitor execution
-sh git-monitor.sh &
-GIT_MONITOR_PID=$!
+	# Start git-monitor execution
+	sh git-monitor.sh &
+	GIT_MONITOR_PID=$!
+else
+	echo -e "No GitHub Repo was provided. Skipping git monitor..."
+fi
 
 java -Xmx$JVM_RAM -Xms$JVM_RAM -XX:+AlwaysPreTouch -XX:+DisableExplicitGC -XX:+ParallelRefProcEnabled -XX:+PerfDisableSharedMem -XX:+UnlockExperimentalVMOptions -XX:+UseG1GC -XX:G1HeapRegionSize=8M -XX:G1HeapWastePercent=5 -XX:G1MaxNewSizePercent=40 -XX:G1MixedGCCountTarget=4 -XX:G1MixedGCLiveThresholdPercent=90 -XX:G1NewSizePercent=30 -XX:G1RSetUpdatingPauseTimePercent=5 -XX:G1ReservePercent=20 -XX:InitiatingHeapOccupancyPercent=15 -XX:MaxGCPauseMillis=200 -XX:MaxTenuringThreshold=1 -XX:SurvivorRatio=32 -Dusing.aikars.flags=https://mcflags.emc.gs -Daikars.new.flags=true -jar server.jar nogui
 
-echo "Minecraft server exited. Killing git monitor.."
-# Kill git-monitor.sh when the Minecraft server exits
-kill $GIT_MONITOR_PID
+if [ -n "${GITHUB_REPO_SSH}" ]; then
+	echo "Minecraft server exited. Killing git monitor.."
+	# Kill git-monitor.sh when the Minecraft server exits
+	kill $GIT_MONITOR_PID
 
-# Wait for git-monitor.sh to clean up/terminate before terminating main script
-wait $GIT_MONITOR_PID
+	# Wait for git-monitor.sh to clean up/terminate before terminating main script
+	wait $GIT_MONITOR_PID
+fi
